+++ src/cljs/cxengage_javascript_sdk/domain/errors.cljs
@@ -968,6 +968,20 @@
    :level "warn"
    :message "Failed to unassign item. No item has been assigned."})
 
+(defn failed-to-assign-salesforce-classic-item-no-interaction-err [interaction-id]
+  {:code 15006
+   :context :salesforce-classic
+   :data {:interaction-id interaction-id}
+   :level "warn"
+   :message "Failed to assign/unassign item. Interaction id does not correspond to active interaction."})
+
+(defn failed-to-assign-blank-salesforce-classic-item-err [interaction-id]
+  {:code 15007
+   :context :salesforce-classic
+   :data {:interaction-id interaction-id}
+   :level "warn"
+   :message "Failed to assign item. Cannot assign blank active tab."})
+
 (defn failed-to-send-salesforce-lightning-assign-err [interaction-id data]
   {:code 16000
    :context :salesforce-lightning